feat: Implement __init__ of DQNAgent
DQNAgent is instantiated with three variables; state_size,
action_size, and seed. It uses two instances of QNetworks.
The local QNetworks is used for the continuous learning of the
agent.
The targe QNetworks is a kind of the approximate to the true
action-value function of the environment.
The learning procedure uses the mean-square-error loss to evaluate
the policy.
The local QNetworks' parameters are updated with an Adam optimizer.
The target QNetworks is updated every UPDATE_EVERY.
ReplayBuffer is stored in an instance variable memory.